代码整改

+++ frameworks/kits/js/common/image_packer_napi.cpp
@@ -147,7 +147,7 @@ STATIC_EXEC_FUNC(Packing)
             bufferSize, context->packOption);
         context->rImagePacker->AddImage(*(context->rImageSource));
         context->rImagePacker->FinalizePacking(packedSize);
-        HiLog::Debug(LABEL, "packedSize=%{public}lld.", static_cast<int64_t>(packedSize));
+        HiLog::Debug(LABEL, "packedSize=%{public}lld.", static_cast<long long>(packedSize));
 
         if (packedSize > 0 && (static_cast<uint64_t>(packedSize) < bufferSize)) {
             context->packedSize = packedSize;
@@ -212,7 +212,7 @@ STATIC_EXEC_FUNC(PackingFromPixelMap)
             bufferSize, context->packOption);
         context->rImagePacker->AddImage(*(context->rPixelMap));
         context->rImagePacker->FinalizePacking(packedSize);
-        HiLog::Debug(LABEL, "packedSize=%{public}lld.", static_cast<int64_t>(packedSize));
+        HiLog::Debug(LABEL, "packedSize=%{public}lld.", static_cast<long long>(packedSize));
 
         if (packedSize > 0 && (static_cast<uint64_t>(packedSize) < bufferSize)) {
             context->packedSize = packedSize;
